Optionally persist userDataDir with preserveUserDataDir flag (#836)
Closes (#825)

# why
some users want to preserve the userDataDir between runs in local (per
https://github.com/browserbase/stagehand/issues/794 and
https://github.com/browserbase/stagehand/pull/825)

# what changed
- new flag in localBrowserLaunchOptions called preserveUserDataDir

# test plan
- added relevant test cases in tests/local/create.test.ts

+++ lib/index.ts
@@ -825,7 +825,10 @@ export class Stagehand {
       }
     }
 
-    if (this.contextPath) {
+    if (
+      this.contextPath &&
+      !this.localBrowserLaunchOptions?.preserveUserDataDir
+    ) {
       try {
         fs.rmSync(this.contextPath, { recursive: true, force: true });
       } catch (e) {
